Etapa 4 - Finalizada com Sucesso (#93)
* [feature] correção create/edituser/userservices

* [feature] CreatedBroadcastList

* [feature] created broadcast list

* [feature] CreatedBroadcastList/CreateUser/EditUser

* feature-191 [Update - Route/Arrow Left/Create Group Fix Bug]

* [ fix ] small fix after merge.

* [ Fix ] Commenting duplicated function

* [ Fix ] fontSize on android

* feature-191 [Update - Header NewPassword]

* [ feature ] Update em: Listagem de lista de transmissão

* [ feature ] Update II em: Listagem de lista de transmissão

* [ feature ] Filtro de quantidade de Listas de transmissão

* [ Fix ] Changing endpoint

* [ Fix ] User context, Image and more

* [ Feature ] E2E for group

* [ feature ] Modificações no Header InactivateUser e UserManagement

* Socket update.

* [ Fix ] E2e decrypt

* [ Fix ] message decrypt from api

* [ feature ] Adicionando destaque aos campos com erro

* [ feature ] Rota após a conclusão da inativação

* [feature] Fixed bug opening profile screen on Android

* feature-192 [Update headers Screens/ChangePassword]

* feature-192 [Clear ScrollView of FlatLists and partials/Console.log]

* [feature] Remove unnecessary imports

* [feature] atualização da develop com a feature-182

* feature-192 [Updates - Scroll Horizontal/URL/otthers]

* feature-192 [Update - UpdateUserPhoto and Navigation ArchivedGroups]]

* [ Feature ] Chat upgrade

* feature-192 [Update setNewPhoto-Profile Screen]

* feature-192 [Update-ChangePhoto Profile Screen]

* [feature] toast, edit/create user, createdbroadcastlist

* [ feature ] Navegação Listas de Transmissão

* [ feature ] Ajustes para o estilo do botão - BroadcastCreate

* [ feature ] Retirando estilo 'bold' do buttonText - InactivateUser

* [ Fix ] Minor changes.

* [ Fix ] context

* [ Fix ] Db not supported

* [feature] Initial group chat

* [feature] Open group without calling API

* [feature] Redirect right from the chat page

* Update 02-Especificação do Projeto.md

* Update 08-Plano de Testes de Software.md

* feature-192 [Update - Profile / InputText]

* Update 09-Registro de Testes de Software.md

* [bug] Fix navigations and edit group remove group admin

* [bug] Fix create group with self as admin and archived groups route

* [bug] Fix only group admin could edit groups

* [bug] Refresh page after edit

* feature-197 [Message - Recuperar Senha na tela Login]

* [bug] Fix chat images not appearing on android

* [ feature ] Botão fixo na parte inferior da tela

* Substituição do useEffect original por useFocusEffect, para que a lista seja recarregada sempre que a tela BroadcastCreate estiver em foco

* [ feature ] Upgrade Scroll

* Update 11-Registro de Testes de Usabilidade.md

Teste de usabilidade 2

* Update 10-Plano de Testes de Usabilidade.md

Alteração nos casos de testes

* Update 11-Registro de Testes de Usabilidade.md

Registro de teste de usabilidade

* Update 09-Registro de Testes de Software.md

Registro de testes de software

* [doc] Adicionado escopo de Programação de Funcionalidades

* feature-199 [Update README.md]

* feature-199 [Update README.md]

* feature-199 [Update 07-Programação de Funcionalidades.md]

* feature-199 [Update README.md]

* [ Fix ] Ajuste para atualizar as informações corretamente ao alterar usuario.

* [doc] Alterações no README.md

* [doc] Alterações no README.md 2

* [ Fix ] Auto scroll ao abrir chat.

* [ doc ] Ajustes na Funcionalidade 05

* [feat] Correções na Homepage

* [update] Toast, CreatedBroadcastList, CreateUser, EditUser

* [feature] Toast

* [ feature ] Estilos para listas arquivadas

* [ feature ] Ajustando estilos dos botões

* [ doc ] Funcionalidade 02

* Update 08-Plano de Testes de Software.md

Alteração do CT 11

* Update 09-Registro de Testes de Software.md

Registro de testes do RF-11 e RNF-06

* [feat] Correções na CreateBroadCastList

* Update 09-Registro de Testes de Software.md

Registro de teste

* Update 11-Registro de Testes de Usabilidade.md

Teste de usabilidade

* Update 11-Registro de Testes de Usabilidade.md

Alterações no teste de usabilidade

* [bug] Ajustar campos superiores sendo escondidos

* [ Fix ] Group & private chat working.

* Update socket.js

* [feature] Ajustados botões da tela de informações de grupo

* [bug] Ajustados campos superiores sendo escondidos

* [feature] att 182 app main bc edituser gc lc

* [ doc ] Update Programação de Funcionalidades

* [bug] Corrigir botão selecione um cargo

* [ doc ] Upgrade Programação de Funcionalidades

* [doc] Atualizar Programação de Funcionalidades.md

* [bug] Padronizar inputs

* [ doc ] Update Template Padrão da Aplicação

* Update 02-Especificação do Projeto.md

Alteração do RNF-06

* Update 08-Plano de Testes de Software.md

Ajuste RNF-06

* Update 09-Registro de Testes de Software.md

Registro de testes RNF-06

* [ doc ] Update Programação de Funcionalidades

* [ doc ] Update Programação de Funcionalidades

* feature 199 [Update 07-Programação de Funcionalidades.md - 03,04,08,09,10]

* [ Feature ] readBy

* [feature] att 182

* [ Feature ] Backend update

* [feature] att edit/create user, delete cbl

* [feature] edituser toast fix

* [feature] fix hompage / groupconversation

* [bug] Remover erro e warning

* [feature] Ajustar estilo de pickers

* [feature] Ajustado para botão ocupar todo o espaço

* [feature] Adicionado caminho para conversa de grupo pela chat

* [doc] Update 07-Programação de Funcionalidades.md

* [doc] Remover lixo de conflitos anteriores

* [doc] Removido lixo de template de programação de funcionalidades

* [doc] Pequenos ajustes em README.md

* [doc] adicionar ao README.md

* [feat] Ajustes Finais

* Update 07-Programação de Funcionalidades.md

* [ Fix ] Group redirect & chat

* [bug] Corrigido bug ao mandar mensagem em grupo

* [bug] Corrigir picker em editar usuário e criar usuário

* [bug] Última versão do chat

* [presentation] videos demonstrando a aplicação em funcionamento  adicionados

* [presentation] video renomeado

* [bug] Corrigido o picker da criar usuário

* [feature] Adicionar teclado do telefone

* [doc] README.md Atualizado

* [doc] Gerenciamento atualizado - Etapa 4

* [doc] Atualizado Gerenciamento do Projeto e Matriz de Rastreabilidade

* [ doc ] Update Programação de Funcionalidades

* [ doc ] Update Template Padrão da Aplicação

---------

+++ src/2gather/contexts/UserContext.js
@@ -1,44 +1,67 @@
 import React, { createContext, useState, useContext, useEffect } from 'react';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { dbGetE2e } from '../services/localDb/user.services';
+import { isAvailableAsync } from 'expo-media-library';
 export const UserContext = createContext();
 
 export default function UserProvider({ children }) {
     const [signed, setSigned] = useState(false);
     const [name, setName] = useState('');
     const [email, setEmail] = useState('');
-    const [telefone, setTelefone] = useState('');
+    const [phone, setPhone] = useState('');
     const [role, setRole] = useState('');
     const [photo, setPhoto] = useState('');
     const [id, setId] = useState('');
-    const [privateE2eContext, setPrivateE2eContextContext] = useState('');
-    const [publicE2eContext, setPublicE2eContextContext] = useState('');
-
+    const [privateE2eContext, setPrivateE2eContext] = useState('');
+    const [publicE2eContext, setPublicE2eContext] = useState('');
+    const [isAdmin, setIsAdmin] = useState(false);
 
     useEffect(() => {
         const fetchData = async () => {
             try {
-                const id = await AsyncStorage.getItem('id');
                 const storedSigned = await AsyncStorage.getItem('signed');
-                const storedName = await AsyncStorage.getItem('name');
-                const access = await AsyncStorage.getItem('access');
-                const refresh = await AsyncStorage.getItem('refresh');
-                const e2eKeys = await dbGetE2e(id)
+                if(storedSigned == 'true') {
+                    setSigned(true)
+                }
+                if(storedSigned == 'true' && signed == true) {
+                    const id = await AsyncStorage.getItem('id');
+                    const storedName = await AsyncStorage.getItem('name');
+                    const access = await AsyncStorage.getItem('access');
+                    const refresh = await AsyncStorage.getItem('refresh');
+                    const photo = await AsyncStorage.getItem('photo');
+                    const email = await AsyncStorage.getItem('email');
+                    const phone = await AsyncStorage.getItem('phone');
+                    const role = await AsyncStorage.getItem('role');
+                    const admin = await AsyncStorage.getItem('isAdmin');
+                    let e2eKeys = null
 
-                // vale fazer um request aqui pra ver se a access/refresh token estão validas, caso n, deslogar.
-                if (storedSigned !== null && access !== null && refresh !== null && id != null) {
-                    setSigned(true);
-                    setId(id);
-                } else {
-                    AsyncStorage.clear()
-                    setSigned(false);
-                }
-                if (storedName !== null) {
+                    try{
+                        e2eKeys = await dbGetE2e(id)
+                    }
+                    catch{}
+
+                    // vale fazer um request aqui pra ver se a access/refresh token estão validas, caso n, deslogar.
+                    if (storedSigned !== null && access !== null && refresh !== null && id != null) {
+                        setSigned(true);
+                        setId(id);
+                    } else {
+                        AsyncStorage.clear()
+                        setSigned(false);
+                    }
                     setName(storedName);
-                }
-                if(e2eKeys){
-                    setPrivateE2eContextContext(e2eKeys.privateKey)
-                    setPublicE2eContextContext(e2eKeys.publicKey)
+                    setPhoto(photo)
+                    setEmail(email)
+                    setPhone(phone)
+                    setRole(role)
+                    setIsAdmin(admin);
+                    if(e2eKeys){
+                        setPrivateE2eContext(e2eKeys.privateKey)
+                        setPublicE2eContext(e2eKeys.publicKey)
+                    }else{
+                        setPrivateE2eContext('KeyNotFound')
+                        setPublicE2eContext('KeyNotFound')
+                    }
+                    
                 }
             } catch (error) {
                 console.error('Error retrieving data from AsyncStorage:', error);
@@ -46,18 +69,6 @@ export default function UserProvider({ children }) {
         };
 
         fetchData();
-    }, []);
-
-    useEffect(() => {
-        const saveData = async () => {
-            try {
-                await AsyncStorage.setItem('signed', JSON.stringify(signed));
-            } catch (error) {
-                console.error('Error saving data to AsyncStorage:', error);
-            }
-        };
-
-        saveData();
     }, [signed]);
 
     return (
@@ -69,40 +80,42 @@ export default function UserProvider({ children }) {
                 setId,
                 name,
                 setName,
+                photo,
+                setPhoto,
+                email,
+                setEmail,
+                phone,
+                setPhone,
+                role,
+                setRole,
                 privateE2eContext,
-                setPrivateE2eContextContext,
+                setPrivateE2eContext,
                 publicE2eContext,
-                setPublicE2eContextContext
+                setPublicE2eContext,
+                isAdmin,
+                setIsAdmin,
             }}>
             {children}
         </UserContext.Provider>
     );
 }
 
-const checkKeys = async() =>{
-    //gambiarra pra logout
-
-    try {
-        const access = await AsyncStorage.getItem('access');
-        const refresh = await AsyncStorage.getItem('refresh');
-    
-        if (!access || !refresh) {
-            const context = useContext(UserContext);
-            const { signed, setSigned, id, setId, name, setName, privateE2eContext, setPrivateE2eContextContext, publicE2eContext, setPublicE2eContextContext } = context;
-            setSigned(false);
-            setId(null);
-            setName(null);
-        }
-      } catch (error) {
-        console.error('Error reading data from AsyncStorage:', error);
-      }
-}
-
 export function useUser() {
-    const context = useContext(UserContext);
-    if(context.signed == true)
-        checkKeys()
+  const context = useContext(UserContext);
 
-    const { signed, setSigned, id, setId, name, setName, privateE2eContext, setPrivateE2eContextContext, publicE2eContext, setPublicE2eContextContext } = context;
-    return { signed, setSigned, id, setId, name, setName, privateE2eContext, setPrivateE2eContextContext, publicE2eContext, setPublicE2eContextContext };
+    const { 
+        signed, setSigned,
+        id, setId,
+        name, setName,
+        photo, setPhoto,
+        email, setEmail,
+        phone, setPhone,
+        role, setRole,
+        privateE2eContext, setPrivateE2eContext,
+        publicE2eContext, setPublicE2eContext,
+        isAdmin,
+        setIsAdmin,
+    } = context;
+
+    return { signed, setSigned, id, setId, name, setName, photo, setPhoto, email, setEmail, phone, setPhone, role, setRole, privateE2eContext, setPrivateE2eContext, publicE2eContext, setPublicE2eContext,  isAdmin, setIsAdmin };
 }
